model-checker : change printf for size_t variables

+++ src/xbt/automaton.c
@@ -343,6 +343,6 @@ int automaton_label_transition_compare(xbt_exp_label_t l1, xbt_exp_label_t l2){
 
 int propositional_symbols_compare_value(const void *s1, const void *s2){
 
-  return (!((int)s1 == (int)s2));
+  return (!(s1 == s2));
 
 }
